Firmware for BL Touch-SILENT BOARD
A basic firmware compiled with the BL Touch enabled using the stock mount that comes with the BL touch.   Use this file in place of the .HEX file provided from Creality when setting up your BL Touch and using the provided USB programmer to flash the board.  After flashing this you will have a bootloader on your board and will be able to update firmware via USB for any future changes.